ai-doc-chatbot/requirements.txt
flask
flask-cors
python-dotenv
pypdf
faiss-cpu
sentence-transformers
groq
openai
google-generativeai
